refactor(agent): move the LLM tool service into internal/agent
Move-only: tools.go implements agentdomain.ToolService over agent/tools'
registry, plus the no-op variant used when tools are disabled. It is the agent
core's own adapter, not a shared service. Renamed to llm_tool_service.go so it
does not read as a second tools package next to agent/tools.
