+++ internal/agent/llm_tool_service.go
@@ -1,4 +1,4 @@
-package services
+package agent
 
 import (
 	"context"
