+++ internal/agent/llm_tool_service_test.go
@@ -1,4 +1,4 @@
-package services
+package agent
 
 import (
 	"slices"
